Rename project from DPAWin to DPAW
Before this project get's used by other people, lets change the project name.
It's still an acronym for Daniel Patrick Abrechts Windowmanager. While
in german (my native language), Windowmanager is one word, in english it's
two words (window manager), so some may have expected the acronym to be DPAWM
instead, since WM is a more commonly used as acronym in other window managers.
However, DPAW was chosen because it's shorter and easier to use in conversation.
Noone is going to spell out the letters in conversations, and D PAW IN or D PAW M
don't flow nicely, while D PAW, sounds like The Paw, which can be used in
conversations without sounding odd.